backend/requirements_docker.txt
requests
beautifulsoup4
flask
flask-cors
python-dotenv
psycopg2-binary
openai
pytest
pytube
youtube-transcript-api
assemblyai
unidecode
boto3
aws-xray-sdk
langfuse
